% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_apply.R
\name{bugfix_openxlsx2_fmt_text}
\alias{bugfix_openxlsx2_fmt_text}
\title{Bugfix for illegal XML chars}
\usage{
bugfix_openxlsx2_fmt_text(x)
}
\arguments{
\item{x}{content with potentially "legal" XML chars}
}
\value{
a character vector
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Only applies for openxlsx versions <= 0.8.0.9000! Otherwise return
x

Adopted from \code{openxlsx2}

For Bug reports, see:
\itemize{
\item \href{https://github.com/pteridin/flexlsx/issues/2}{Bugreport #2}
\item \href{https://github.com/JanMarvin/openxlsx2/pull/757}{Upstream}
}
}
