% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_xlsx.R
\name{compare_xlsx}
\alias{compare_xlsx}
\title{Hashes the contents of xlsx files}
\usage{
compare_xlsx(tmpfile)
}
\arguments{
\item{tmpfile}{the file to be hashed}
}
\value{
A list of hashed values for certain xml files
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
