% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexlsx-package.R
\docType{package}
\name{flexlsx-package}
\alias{flexlsx}
\alias{flexlsx-package}
\title{flexlsx: Exporting 'flextable' to Excel}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Exports 'flextable' tables to Excel-Sheets using 'openxlsx2'. Objects that can be converted to 'flextable' objects (e.g. 'gtsummary') can be exported as well.
}
\author{
\strong{Maintainer}: Tobias Heidler \email{tobias.heidler@googlemail.com} (\href{https://orcid.org/0000-0001-9193-0980}{ORCID}) [copyright holder]

}
\keyword{internal}
