% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_apply_border.R
\name{ft_to_xlsx_border_width}
\alias{ft_to_xlsx_border_width}
\title{Determines the border width}
\usage{
ft_to_xlsx_border_width(border_width)
}
\arguments{
\item{border_width}{a numeric vector determining the border-width}
}
\value{
a factor of xlsx border styles
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
