% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_to_style.R
\name{ftpart_to_style_tibble}
\alias{ftpart_to_style_tibble}
\title{Converts a flextable-part to a tibble styles}
\usage{
ftpart_to_style_tibble(ft_part, part = c("header", "body", "footer"))
}
\arguments{
\item{ftpart}{a \link[flextable:flextable-package]{flextable} part}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
flextable::flextable(mtcars) -> ft
ftpart_to_tibble(ft$header, "header")

}
