% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_ranges.R
\name{get_dim_colwise}
\alias{get_dim_colwise}
\title{Groups each row with same style each column}
\usage{
get_dim_colwise(df_rows)
}
\arguments{
\item{df_rows}{\link[tibble:tibble-package]{tibble} of row-wise aggregates style}
}
\value{
\link[tibble:tibble-package]{tibble} of column-wise aggregates style
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
