% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_ranges.R
\name{get_dim_ranges}
\alias{get_dim_ranges}
\title{Retrieves dims of same style rows within same column}
\usage{
get_dim_ranges(df_x)
}
\arguments{
\item{df_x}{styling information incl. col_id & row_id}
}
\value{
merged styles as a \link[tibble:tibble-package]{tibble}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
