% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_ranges.R
\name{get_dim_rowwise}
\alias{get_dim_rowwise}
\title{Groups each column with same style each row}
\usage{
get_dim_rowwise(df_x, df_style_hashed)
}
\arguments{
\item{df_x}{styling information incl. col_id & row_id}

\item{df_style_hashed}{\link[tibble:tibble-package]{tibble} of hashed style
information}
}
\value{
\link[tibble:tibble-package]{tibble} of row-wise aggregates style
information
}
\description{
Groups each column with same style each row
}
