% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_apply_border.R
\name{handle_null_border}
\alias{handle_null_border}
\title{Where there is no border return NULL}
\usage{
handle_null_border(border_width)
}
\arguments{
\item{border_width}{a numeric vector determining the border-width}
}
\value{
border_width or NULL
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
