% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_apply_merge.R
\name{merge_resolve_type}
\alias{merge_resolve_type}
\title{Determine problematic merges}
\usage{
merge_resolve_type(df_to_merge)
}
\arguments{
\item{df_to_merge}{The data.frame containing information about the cells to merge}
}
\value{
df_to_merge is extended by is_encapsulated and is_need_resolve
}
\description{
Determine problematic merges
}
