% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prepare_color}
\alias{prepare_color}
\title{Prepares the color for content style}
\usage{
prepare_color(color_name)
}
\arguments{
\item{color_name}{The name of the color}
}
\value{
The hexadecimal RGB-value
}
\description{
Converts a color name to the hexadecimal RGB-value
Removes "transparent" color
}
