% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_ranges.R
\name{style_to_hash}
\alias{style_to_hash}
\title{Retrieves hashed style information}
\usage{
style_to_hash(df_x)
}
\arguments{
\item{df_x}{styling information incl. col_id & row_id}
}
\value{
hashed style information as a \link[tibble:tibble-package]{tibble}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Converts each style to an individual integer hash
for easy comparison and aggregation.
}
