% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_add_caption.R
\name{wb_add_caption}
\alias{wb_add_caption}
\title{Adds a caption to an excel file}
\usage{
wb_add_caption(
  wb,
  sheet,
  ft,
  offset_rows = offset_rows,
  offset_cols = offset_cols
)
}
\arguments{
\item{wb}{an openxlsx2 workbook}

\item{sheet}{an openxlsx2 workbook sheet}

\item{ft}{a flextable}

\item{offset_rows}{zero-based row offset}

\item{offset_cols}{zero-based column offset}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
