% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_apply.R
\name{wb_add_flextable}
\alias{wb_add_flextable}
\title{Adds a flextable to an openxlsx2 workbook sheet}
\usage{
wb_add_flextable(wb, sheet, ft)
}
\arguments{
\item{wb}{an openxlsx2 workbook}

\item{sheet}{an openxlsx2 workbook sheet}

\item{ft}{a flextable}
}
\description{
Adds a flextable to an openxlsx2 workbook sheet
}
\examples{
ft <- flextable::as_flextable(table(mtcars[,1:2]))
wb <- openxlsx2::wb_workbook()$add_worksheet("mtcars")
wb_add_flextable(wb, "mtcars", ft)$save("~/text.xlsx")
}
