% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_apply.R
\name{wb_apply_cell_styles}
\alias{wb_apply_cell_styles}
\title{Applies the cell styles}
\usage{
wb_apply_cell_styles(wb, sheet, df_style)
}
\arguments{
\item{wb}{the \link[openxlsx2:wbWorkbook]{workbook}}

\item{sheet}{the sheet of the workbook}

\item{df_style}{the styling tibble from \link{ft_to_style_tibble}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
ft <- flextable::as_flextable(table(mtcars[,1:2]))
df_style <- ft_to_style_tibble(ft)
wb <- openxlsx2::wb_workbook()$add_worksheet("mtcars")
wb_apply_cell_styles(wb, "mtcars", df_style)
}
