% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_apply_merge.R
\name{wb_apply_merge}
\alias{wb_apply_merge}
\title{Merges cells}
\usage{
wb_apply_merge(wb, sheet, df_style)
}
\arguments{
\item{wb}{the \link[openxlsx2:wbWorkbook]{workbook}}

\item{sheet}{the sheet of the workbook}

\item{df_style}{the styling tibble from \link{ft_to_style_tibble}}
}
\value{
df_style tibble
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
