% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_change_height_width.R
\name{wb_change_cell_width}
\alias{wb_change_cell_width}
\title{Changes the cell width}
\usage{
wb_change_cell_width(wb, sheet, ft, offset_cols)
}
\arguments{
\item{wb}{an openxlsx2 workbook}

\item{sheet}{an openxlsx2 workbook sheet}

\item{ft}{a flextable}

\item{offset_cols}{zero-based column offset}
}
\description{
Changes the cell width
}
