% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_change_height_width.R
\name{wb_change_row_height}
\alias{wb_change_row_height}
\title{Changes the row height}
\usage{
wb_change_row_height(wb, sheet, df_style)
}
\arguments{
\item{wb}{an openxlsx2 workbook}

\item{sheet}{an openxlsx2 workbook sheet}

\item{df_style}{the styling tibble from \link{ft_to_style_tibble}}
}
\description{
Changes the row height
}
