% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexlsx-package.R
\docType{package}
\name{flexlsx-package}
\alias{flexlsx}
\alias{flexlsx-package}
\title{flexlsx: Exporting Flextables to Excel}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Exports flextables to Excel-Sheets using openxlsx2. Objects that can be converted to flextables (e.g. gtsummary) can be exported as well.
}
\author{
\strong{Maintainer}: Tobias Heidler \email{tobias.heidler@googlemail.com} (\href{https://orcid.org/0000-0001-9193-0980}{ORCID})

}
\keyword{internal}
