% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexlsx-package.R
\docType{package}
\name{flexlsx-package}
\alias{flexlsx}
\alias{flexlsx-package}
\title{flexlsx: Exporting 'flextable' to 'xlsx' Files}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Exports 'flextable' objects to 'xlsx' files, utilizing functionalities provided by 'flextable' and 'openxlsx2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pteridin/flexlsx}
  \item Report bugs at \url{https://github.com/pteridin/flexlsx/issues}
}

}
\author{
\strong{Maintainer}: Tobias Heidler \email{tobias.heidler@googlemail.com} (\href{https://orcid.org/0000-0001-9193-0980}{ORCID}) [copyright holder]

}
\keyword{internal}
