% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_to_style.R
\name{ft_to_style_tibble}
\alias{ft_to_style_tibble}
\title{Converts a flextable to a tibble with style information}
\usage{
ft_to_style_tibble(
  ft,
  offset_rows = 0L,
  offset_cols = 0L,
  offset_caption_rows = 0L
)
}
\arguments{
\item{ft}{a \link[flextable:flextable-package]{flextable}}

\item{offset_rows}{offsets the start-row}

\item{offset_cols}{offsets the start-columns}

\item{offset_caption_rows}{number of rows to offset the caption by}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
